requirements.txt
Django~=5.0.4
python-dotenv~=1.0.1
aiogram~=3.4.1
openai~=1.35
asgiref~=3.8.1
scipy~=1.14.1
tiktoken~=0.8.0